Main and Recipe class updated
Constructor has been updated and now stores the ingredient quantity and ingredient unit of measurement. Main class is now updated and now asks for a certain amount of ingredients to be typed including ingredient name, quantity and unit of measurement which is then stored in their own array and their own array is then stored in the constructor. toString is updated to allow for ingredients info to be displayed. Recipe steps yet to be done