Implemented steps to take for recipe
Constructor now also stores the steps needed to make recipe in the Recipe class and the Main class has been updated to allow for users to enter the number of steps for recipe as well as typing down the recipes. The steps will be safely stored in an array then that array for steps will be stored in the constructor. The toString method has been updated to allow the steps taken to make recipe to also be displayed. Scaling recipe by factor feature is yet to be implemented.